requirements.txt
labjack-ljm==1.21.0
msgpack==1.0.7
numpy==1.26.0
orjson==3.9.7
packaging==23.2
pandas==2.1.1
python-can==4.2.2
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
typing_extensions==4.8.0
tzdata==2023.3
wrapt==1.15.0
